Conceptually log dequeue in AOF.
This is a big semantical change from the original Disque. Now we log
dequeue events, so that by default we just load the jobs data on restart
and put things back in the queue if the job state is "queued". If we want to
be safe against re-deliveries of at-most-once jobs (since the AOF may not
always be in sync), there is a specific option that the user can use to
configure Disque in order to start with the jobs not in queue, in case
their "retry" time is 0.