DISQUE INFO: emit all fields as an array.
RESP3 can have verbatim output, but it's going to take a while. Who uses
RESP2 would see just a string with all the garbage of all the fields.
Putting non Redis native stuff as exceptions inside redis-cli is not
viable. So what's the solution? Emit it as an array of strings.